requirements.txt
streamlit==1.32.0
openai-whisper==20231117
pyannote.audio==3.1.1
langchain==0.1.13
langchain-groq==0.1.0
python-dotenv==1.0.1
python-docx==1.1.0
fpdf2==2.7.8
torchaudio==2.2.1
